Split Swagger documentation for V1 and V2
