+++ src/main/java/com/bmw/remotecollab/admin/config/SwaggerConfig.java
@@ -15,15 +15,27 @@ import springfox.documentation.swagger2.annotations.EnableSwagger2;
 public class SwaggerConfig {
 
     @Bean
-    public Docket api() {
+    public Docket apiV1() {
         return new Docket(DocumentationType.SWAGGER_2)
+                .groupName("API Version 1")
                 .select()
                 .apis(RequestHandlerSelectors
                         .basePackage("com.bmw.remotecollab"))
-                .paths(PathSelectors.any())
+                .paths(PathSelectors.ant("/api/v1/**"))
                 .build()
-                .apiInfo(apiInfo())
-        ;
+                .apiInfo(apiInfo());
+    }
+
+    @Bean
+    public Docket apiV2() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .groupName("API Version 2")
+                .select()
+                .apis(RequestHandlerSelectors
+                        .basePackage("com.bmw.remotecollab"))
+                .paths(PathSelectors.ant("/api/v2/**"))
+                .build()
+                .apiInfo(apiInfo());
     }
 
     private ApiInfo apiInfo() {
