+++ src/main/java/com/bmw/remotecollab/admin/rest/v1/requests/RequestJoinRoom.java
@@ -1,6 +1,5 @@
 package com.bmw.remotecollab.admin.rest.v1.requests;
 
-import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -9,7 +8,6 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-@ApiModel("RequestJoinRoomV1")
 public class RequestJoinRoom {
 
     @ApiModelProperty(value = "The UUID of the room, you want to join.", required = true)
