+++ src/main/java/com/bmw/remotecollab/admin/rest/v1/requests/RequestNewRoom.java
@@ -1,7 +1,6 @@
 package com.bmw.remotecollab.admin.rest.v1.requests;
 
 import com.bmw.remotecollab.admin.service.email.EmailList;
-import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -13,7 +12,6 @@ import java.util.List;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-@ApiModel("RequestNewRoomV2")
 public class RequestNewRoom {
 
     @ApiModelProperty(value = "The name of the room, you want to create.", required = true)
