+++ src/main/java/com/bmw/remotecollab/admin/rest/v2/requests/RequestInviteUser.java
@@ -1,7 +1,6 @@
 package com.bmw.remotecollab.admin.rest.v2.requests;
 
 import com.bmw.remotecollab.admin.service.email.EmailList;
-import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -12,7 +11,6 @@ import java.util.List;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-@ApiModel("RequestInviteUserV2")
 public class RequestInviteUser {
 
     @ApiModelProperty(value = "A non-null, non-empty list of valid email adresses.", required = true)
