+++ src/main/java/com/bmw/remotecollab/admin/rest/v2/responses/ResponseNewRoom.java
@@ -1,6 +1,5 @@
 package com.bmw.remotecollab.admin.rest.v2.responses;
 
-import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -11,7 +10,6 @@ import java.util.Date;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-@ApiModel("ResponseNewRoomV2")
 public class ResponseNewRoom {
 
     @ApiModelProperty(value = "The UUID of the room, you created.")
